requirements.txt
hdx-python-api==6.0.5
requests==2.31.0
